+++ R/get_fields.R
@@ -18,3 +18,25 @@ get_fields <- function() {
   dplyr::bind_cols(dplyr::select(all_fields, -.data$schema),
                    purrr::pluck(all_fields, "schema"))
 }
+
+#' Update the internal mapping of the field containing story points
+#'
+#' The story point field in Jira unfortunately is mapped to a random
+#' custom field identifier in each instance/project. `find_story_point_mapping()`
+#' queries the field list and updates the `sprintr` mapping. This value is
+#' not persisted and needs to be re-run each session.
+#'
+#' @return Invisibly returns the new value of the story points field
+#' @export
+#'
+#' @importFrom glue glue
+#'
+#' @examples
+#' NULL
+find_story_point_mapping <- function() {
+  fields <- get_fields()
+  sp_field <- fields[fields$name == "Story Points", "id"]
+  message(glue::glue("Mapping story_points to {sp_field}."))
+  pkg.globals$story_points <- sp_field
+  return(invisible(sp_field))
+}