+++ R/get_issue.R
@@ -29,11 +29,11 @@ get_issue <- function(issue_key, fields = NULL, full_response = FALSE) {
   if (is.null(fields) && !full_response) {
     tibble::tibble(
       key = issue_key,
-      story_points = purrr::pluck(resp, "customfield_10013",
+      story_points = purrr::pluck(resp, pkg.globals$story_points,
                                   .default = NA_integer_),
       epic_name = purrr::pluck(resp, "epic", "name",
                                .default = NA_character_),
-      program = purrr::pluck(resp, "customfield_10500", "value",
+      program = purrr::pluck(resp, pkg.globals$program, "value",
                              .default = NA_character_)
     )
   } else {